+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 
 ## Unreleased
+### Changed
+- Require PHP 7.1+
+
+### Removed
+- `TestUtils` class which was already deprecated in 2.1.
+    - Instead of `TestUtils::setSelect2Value()` use directly the new `Select2` component and `selectByVisiblePartialText()` method.
+    - Instead of `TestUtils::getFixturePath()` use `Facebook\WebDriver\Remote\FileDetector` instead.
+    - Instead of `TestUtils::sleep()` use `AbstractTestCase::sleep()` method instead.
+- `getConfig()` method of ConfigProvider. Instead call the property directly on instance of the ConfigProvider.
 
 ## 2.3.2 - 2017-12-02
 ### Changed